refactor(mvk): carve nucleus-ifc-kernel crate out of portcullis-core (M3)
Physically split the IFC admission core into its own crate so the verified
reference monitor is a thing Cargo enforces, not just a text-ratchet boundary.

nucleus-ifc-kernel now owns: the IFC label lattice (ifc_lattice), the
operation/sink vocabulary (ifc_ops), CapabilityLevel (capability_level), effect,
storage_lane, the FlowTracker + exfiltration-safety check (flow, ifc_api), the
policy-discharge pipeline (discharge), and the Aeneas-extracted IFC slices
(extracted). discharge + CapabilityLevel move WITH the cluster, which dissolves
the ifc_api->discharge cycle (the kernel names no downstream crate).

portcullis-core depends on the kernel and re-exports its full surface
(`pub use nucleus_ifc_kernel::*` + module re-exports), so every existing
`portcullis_core::{IFCLabel, Operation, FlowTracker, discharge::*, extracted::*}`
path is unchanged — verified: portcullis-core + dependents (nucleus-ifc, ...)
build via the re-export, no cycle. The boundary ratchet moved into the kernel
crate (kernel tests: 163 + ratchet green).

Proof-extraction retarget (CI-verified — Charon/Aeneas/Lean are CI-only):
aeneas-ifc-scoped.yml + the IFC theorem files + generated-ifc namespaces
retargeted portcullis_core::extracted -> nucleus_ifc_kernel::extracted.

KNOWN-FOLLOW-UP (next commit, CI will be red until then): moving CapabilityLevel
+ is_exfil_operation out of portcullis-core fractures two single-crate Charon
extractions — aeneas.yml (whole-crate capability) and aeneas-decide-pure.yml
(decide_pure now calls is_exfil cross-crate). These need a deliberate re-point,
addressed next.
