+++ .github/workflows/aeneas-ifc-scoped.yml
@@ -1,8 +1,10 @@
 name: Aeneas IFC (scoped extraction + noninterference)
 
 # Scoped Charon→Aeneas extraction of the integrity-axis enforcement slice
-# (crates/portcullis-core/src/extracted/ifc_integrity.rs) plus the Lean
-# noninterference theorem proven OVER the generated defs.
+# (crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs) plus the Lean
+# noninterference theorem proven OVER the generated defs. (The IFC source
+# carved out of portcullis-core into the dependency-free nucleus-ifc-kernel
+# crate in MVK M3; the Lean project itself stays in portcullis-core/lean.)
 #
 # WHY A SEPARATE WORKFLOW (not aeneas.yml):
 #   aeneas.yml runs `charon cargo --preset aeneas` whole-crate, then
@@ -20,7 +22,7 @@ on:
   push:
     branches: ["main"]
     paths:
-      - "crates/portcullis-core/src/extracted/**"
+      - "crates/nucleus-ifc-kernel/src/extracted/**"
       - "crates/portcullis-core/lean/generated-ifc/**"
       - "crates/portcullis-core/lean/IntegrityNoninterferenceExtracted.lean"
       - "crates/portcullis-core/lean/ConfidentialityNoninterferenceExtracted.lean"
@@ -28,7 +30,7 @@ on:
       - ".github/workflows/aeneas-ifc-scoped.yml"
   pull_request:
     paths:
-      - "crates/portcullis-core/src/extracted/**"
+      - "crates/nucleus-ifc-kernel/src/extracted/**"
       - "crates/portcullis-core/lean/generated-ifc/**"
       - "crates/portcullis-core/lean/IntegrityNoninterferenceExtracted.lean"
       - "crates/portcullis-core/lean/ConfidentialityNoninterferenceExtracted.lean"
@@ -47,7 +49,7 @@ env:
   # Charon needs a specific nightly for rustc-dev components.
   CHARON_NIGHTLY: nightly-2026-06-01
   # The scoped extraction roots — name them exactly as Charon sees the paths.
-  EXTRACT_ROOTS: "portcullis_core::extracted::ifc_integrity::iflows_to,portcullis_core::extracted::ifc_integrity::imeet,portcullis_core::extracted::ifc_integrity::irun_step,portcullis_core::extracted::ifc_confidentiality::cflows_to,portcullis_core::extracted::ifc_confidentiality::cjoin,portcullis_core::extracted::ifc_confidentiality::crun_step"
+  EXTRACT_ROOTS: "nucleus_ifc_kernel::extracted::ifc_integrity::iflows_to,nucleus_ifc_kernel::extracted::ifc_integrity::imeet,nucleus_ifc_kernel::extracted::ifc_integrity::irun_step,nucleus_ifc_kernel::extracted::ifc_confidentiality::cflows_to,nucleus_ifc_kernel::extracted::ifc_confidentiality::cjoin,nucleus_ifc_kernel::extracted::ifc_confidentiality::crun_step"
 
 jobs:
   aeneas-ifc-scoped:
@@ -68,7 +70,7 @@ jobs:
       - name: Rust / cargo cache (Charon target dir)
         uses: Swatinem/rust-cache@42dc69e1aa15d09112580998cf2ef0119e2e91ae # v2.9.1
         with:
-          workspaces: crates/portcullis-core
+          workspaces: crates/nucleus-ifc-kernel
           key: charon-ifc-${{ env.CHARON_NIGHTLY }}-${{ env.AENEAS_RELEASE }}
           save-if: ${{ github.ref == 'refs/heads/main' }}
 
@@ -82,7 +84,7 @@ jobs:
           toolchain: stable
 
       - name: Parity tests (extracted slice == production enforcement)
-        run: cargo test -p portcullis-core --lib extracted
+        run: cargo test -p nucleus-ifc-kernel --lib extracted
 
       # --- Aeneas prebuilt binaries (Charon + Aeneas, no OCaml build) ---
       - name: Cache Aeneas binary
@@ -116,7 +118,7 @@ jobs:
       # --start-from each root walks ONLY the reachable callees, so the
       # crashing unrelated function bodies are never translated by Aeneas.
       - name: Charon scoped MIR extraction
-        working-directory: crates/portcullis-core
+        working-directory: crates/nucleus-ifc-kernel
         run: |
           STARTS=""
           IFS=',' read -ra ROOTS <<< "$EXTRACT_ROOTS"
@@ -130,7 +132,12 @@ jobs:
       - name: Aeneas Lean 4 translation (scoped slice)
         run: |
           mkdir -p crates/portcullis-core/lean/ci-generated-ifc
-          aeneas -backend lean -split-files portcullis_core.llbc \
+          # Charon now runs in crates/nucleus-ifc-kernel (the IFC source crate
+          # moved out of portcullis-core in MVK M3), so locate the emitted
+          # nucleus_ifc_kernel.llbc whether it lands in the crate dir or repo root.
+          LLBC="$(find crates/nucleus-ifc-kernel . -maxdepth 2 -name 'nucleus_ifc_kernel.llbc' -print -quit)"
+          test -n "$LLBC" || { echo "ERROR: charon produced no nucleus_ifc_kernel.llbc"; exit 1; }
+          aeneas -backend lean -split-files "$LLBC" \
             -dest crates/portcullis-core/lean/ci-generated-ifc
 
       # --- Canonicalize: build the theorem against the FRESH extraction ---
@@ -141,7 +148,11 @@ jobs:
       # warning), NOT a validity failure — the committed copy is genuine Aeneas
       # output too; we simply prefer the current toolchain's output. The ONE
       # disclosed hand-edit is retargeting the Funs.lean import
-      # (PortcullisCore.Types -> PortcullisCoreIFC.Types).
+      # (NucleusIfcKernel.Types -> PortcullisCoreIFC.Types). Aeneas derives the
+      # Lean module name from the crate (nucleus_ifc_kernel -> NucleusIfcKernel)
+      # since the IFC source moved to crates/nucleus-ifc-kernel in MVK M3; the
+      # committed Lean lib keeps the stable PortcullisCoreIFC module name so it
+      # does not collide with the whole-crate «PortcullisCore» lib.
       - name: Canonicalize generated-ifc from the fresh extraction
         run: |
           DST=crates/portcullis-core/lean/generated-ifc/PortcullisCoreIFC
@@ -155,7 +166,7 @@ jobs:
             fi
           done
           cp "$SRC/Types.lean" "$DST/Types.lean"
-          sed 's/PortcullisCore\.Types/PortcullisCoreIFC.Types/' "$SRC/Funs.lean" > "$DST/Funs.lean"
+          sed 's/NucleusIfcKernel\.Types/PortcullisCoreIFC.Types/' "$SRC/Funs.lean" > "$DST/Funs.lean"
           echo "generated-ifc/ canonicalized from the fresh extraction; theorem builds against it."
 
       - name: Dump FULL canonical generated defs (to develop the proof against)
