+++ crates/nucleus-ifc-kernel/src/extracted/ifc_confidentiality.rs
