+++ crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs
