+++ crates/nucleus-ifc-kernel/src/flow.rs
@@ -95,7 +95,7 @@ pub enum NodeKind {
     /// to the most conservative label (adversarial integrity, public conf).
     ///
     /// ```rust
-    /// use portcullis_core::flow::NodeKind;
+    /// use nucleus_ifc_kernel::flow::NodeKind;
     ///
     /// let custom = NodeKind::Custom("slack_message");
     /// ```
@@ -1067,8 +1067,8 @@ mod kani_proofs {
 /// # Usage
 ///
 /// ```rust
-/// use portcullis_core::flow::{ZkFlowInput, FlowNode, FlowVerdict, NodeKind, MAX_PARENTS};
-/// use portcullis_core::IFCLabel;
+/// use nucleus_ifc_kernel::flow::{ZkFlowInput, FlowNode, FlowVerdict, NodeKind, MAX_PARENTS};
+/// use nucleus_ifc_kernel::IFCLabel;
 ///
 /// let node = FlowNode {
 ///     id: 1,
@@ -1133,8 +1133,8 @@ pub enum VerificationResult {
 /// (no I/O, no allocation beyond the input) and deterministic.
 ///
 /// ```rust
-/// use portcullis_core::flow::*;
-/// use portcullis_core::*;
+/// use nucleus_ifc_kernel::flow::*;
+/// use nucleus_ifc_kernel::*;
 ///
 /// let source = FlowNode {
 ///     id: 1, kind: NodeKind::FileRead,
