+++ crates/nucleus-ifc-kernel/src/ifc_lattice.rs
