+++ crates/nucleus-ifc-kernel/src/ifc_ops.rs
