+++ crates/portcullis-core/lean/generated-ifc/PortcullisCoreIFC/Funs.lean
@@ -1,5 +1,5 @@
 -- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
--- [portcullis_core]: function definitions
+-- [nucleus_ifc_kernel]: function definitions
 import Aeneas
 import PortcullisCoreIFC.Types
 open Aeneas Aeneas.Std Result ControlFlow Error
@@ -13,10 +13,10 @@ set_option maxHeartbeats 1000000
 /- You can set the `maxRecDepth` value with the `-max-recdepth` CLI option -/
 set_option maxRecDepth 2048
 
-namespace portcullis_core
+namespace nucleus_ifc_kernel
 
-/-- [portcullis_core::extracted::ifc_integrity::irank]:
-    Source: 'crates/portcullis-core/src/extracted/ifc_integrity.rs', lines 74:0-80:1
+/-- [nucleus_ifc_kernel::extracted::ifc_integrity::irank]:
+    Source: 'crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs', lines 74:0-80:1
     Visibility: public -/
 def extracted.ifc_integrity.irank
   (l : extracted.ifc_integrity.IntegLevel) : Result Std.U8 := do
@@ -25,8 +25,8 @@ def extracted.ifc_integrity.irank
   | extracted.ifc_integrity.IntegLevel.Untrusted => ok 1#u8
   | extracted.ifc_integrity.IntegLevel.Trusted => ok 2#u8
 
-/-- [portcullis_core::extracted::ifc_integrity::imeet]:
-    Source: 'crates/portcullis-core/src/extracted/ifc_integrity.rs', lines 89:0-95:1
+/-- [nucleus_ifc_kernel::extracted::ifc_integrity::imeet]:
+    Source: 'crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs', lines 89:0-95:1
     Visibility: public -/
 def extracted.ifc_integrity.imeet
   (a : extracted.ifc_integrity.IntegLevel)
@@ -39,8 +39,8 @@ def extracted.ifc_integrity.imeet
   then ok a
   else ok b
 
-/-- [portcullis_core::extracted::ifc_integrity::iflows_to]:
-    Source: 'crates/portcullis-core/src/extracted/ifc_integrity.rs', lines 103:0-105:1
+/-- [nucleus_ifc_kernel::extracted::ifc_integrity::iflows_to]:
+    Source: 'crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs', lines 103:0-105:1
     Visibility: public -/
 def extracted.ifc_integrity.iflows_to
   (a : extracted.ifc_integrity.IntegLevel)
@@ -51,8 +51,8 @@ def extracted.ifc_integrity.iflows_to
   let i1 ← extracted.ifc_integrity.irank ceiling
   ok (i >= i1)
 
-/-- [portcullis_core::extracted::ifc_integrity::irun_step]:
-    Source: 'crates/portcullis-core/src/extracted/ifc_integrity.rs', lines 110:0-112:1
+/-- [nucleus_ifc_kernel::extracted::ifc_integrity::irun_step]:
+    Source: 'crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs', lines 110:0-112:1
     Visibility: public -/
 def extracted.ifc_integrity.irun_step
   (eff : extracted.ifc_integrity.IntegLevel)
@@ -61,4 +61,4 @@ def extracted.ifc_integrity.irun_step
   := do
   extracted.ifc_integrity.imeet eff src
 
-end portcullis_core
+end nucleus_ifc_kernel
