+++ crates/portcullis-core/lean/generated-ifc/PortcullisCoreIFC/Types.lean
@@ -1,5 +1,5 @@
 -- THIS FILE WAS AUTOMATICALLY GENERATED BY AENEAS
--- [portcullis_core]: type definitions
+-- [nucleus_ifc_kernel]: type definitions
 import Aeneas
 open Aeneas Aeneas.Std Result ControlFlow Error
 set_option linter.dupNamespace false
@@ -12,10 +12,10 @@ set_option maxHeartbeats 1000000
 /- You can set the `maxRecDepth` value with the `-max-recdepth` CLI option -/
 set_option maxRecDepth 2048
 
-namespace portcullis_core
+namespace nucleus_ifc_kernel
 
-/-- [portcullis_core::extracted::ifc_integrity::IntegLevel]
-    Source: 'crates/portcullis-core/src/extracted/ifc_integrity.rs', lines 62:0-69:1
+/-- [nucleus_ifc_kernel::extracted::ifc_integrity::IntegLevel]
+    Source: 'crates/nucleus-ifc-kernel/src/extracted/ifc_integrity.rs', lines 62:0-69:1
     Visibility: public -/
 @[discriminant u8]
 inductive extracted.ifc_integrity.IntegLevel where
@@ -23,4 +23,4 @@ inductive extracted.ifc_integrity.IntegLevel where
 | Untrusted : extracted.ifc_integrity.IntegLevel
 | Trusted : extracted.ifc_integrity.IntegLevel
 
-end portcullis_core
+end nucleus_ifc_kernel
